+++ STACpopulator/populator_base.py
@@ -16,6 +16,7 @@ from STACpopulator.api_requests import (
     stac_host_reachable,
     stac_version_match,
 )
+from STACpopulator.collection_update import UpdateModes, update_collection
 from STACpopulator.input import GenericLoader
 from STACpopulator.models import AnyGeometry
 from STACpopulator.stac_utils import load_config
@@ -33,7 +34,7 @@ class STACpopulatorBase(ABC):
         update: bool = False,
         session: Optional[Session] = None,
         config_file: Optional[Union[os.PathLike[str], str]] = "collection_config.yml",
-        update_collection: Literal["extents", "summaries", "all", "none"] = "none",
+        update_collection: UpdateModes | Literal["none"] = "none",
         exclude_summaries: Iterable[str] = (),
     ) -> None:
         """Initialize the STAC populator.
@@ -53,10 +54,7 @@ class STACpopulatorBase(ABC):
         self._stac_host = self.validate_host(stac_host)
         self.update = update
         self.update_collection = update_collection
-
-        # the STAC spec does not recommend repeating summaries that are covered by the extent already
-        self.exclude_summaries = ["datetime", "start_datetime", "end_datetime"]
-        self.exclude_summaries.extend(exclude_summaries)
+        self.exclude_summaries = exclude_summaries
 
         LOGGER.info("Initialization complete")
         LOGGER.info(f"Collection {self.collection_name} is assigned ID {self.collection_id}")
@@ -122,9 +120,6 @@ class STACpopulatorBase(ABC):
 
         return stac_host
 
-    # FIXME: should provide a way to update after item generation
-    #   STAC collections are supposed to include 'summaries' with
-    #   an aggregation of all supported 'properties' by its child items
     @functools.cache
     def create_stac_collection(self) -> dict[str, Any]:
         """
@@ -189,149 +184,9 @@ class STACpopulatorBase(ABC):
         """Publish this collection by uploading it to the STAC catalog at self.stac_host."""
         post_stac_collection(self.stac_host, collection_data, self.update, session=self._session)
 
-    @staticmethod
-    def _check_wgs84_compliance(bbox: list[int | float], stac_object_type: str, stac_object_id: str | None) -> None:
-        longitude = (bbox[0], bbox[len(bbox) // 2])
-        latitude = (bbox[1], bbox[(len(bbox) // 2) + 1])
-        for lon in longitude:
-            if lon < -180 or lon > 180:
-                LOGGER.warning(
-                    "STAC %s with id [%s] contains a bbox with a longitude outside of the accepted range of -180 and 180",
-                    stac_object_type,
-                    stac_object_id,
-                )
-        for lat in latitude:
-            if lat < -90 or lat > 90:
-                LOGGER.warning(
-                    "STAC %s with id [%s] contains a bbox with a latitude outside of the accepted range of -90 and 90",
-                    stac_object_type,
-                    stac_object_id,
-                )
-
-    @staticmethod
-    def _sorted_bbox(bbox: list[int | float]) -> list[int | float]:
-        return [
-            a for b in zip(*(sorted(axis) for axis in zip(bbox[: len(bbox) // 2], bbox[len(bbox) // 2 :]))) for a in b
-        ]
-
-    def _update_collection_bbox(self, stac_item: dict[str, Any]) -> None:
-        item_bbox = stac_item.get("bbox")
-        if item_bbox is None:
-            # bbox can be missing if there is no geometry
-            return
-        sorted_bbox = self._sorted_bbox(item_bbox)
-        if item_bbox != sorted_bbox:
-            LOGGER.warning(
-                "STAC item with id [%s] contains a bbox with unsorted values: %s should be %s",
-                stac_item.get("id"),
-                item_bbox,
-                sorted_bbox,
-            )
-            item_bbox = sorted_bbox
-        self._check_wgs84_compliance(item_bbox, "item", stac_item.get("id"))
-        collection_bboxes = self._collection["extent"]["spatial"]["bbox"]
-        if collection_bboxes:
-            collection_bbox = collection_bboxes[0]
-            if len(item_bbox) == 4 and len(collection_bbox) == 6:
-                # collection bbox has a z axis and item bbox does not
-                item_bbox = [*item_bbox[:2], collection_bbox[2], item_bbox[2:], collection_bbox[5]]
-            elif len(item_bbox) == 6 and len(collection_bbox) == 4:
-                # item bbox has a z axis and collection bbox does not
-                collection_bbox.insert(2, item_bbox[2])
-                collection_bbox.append(item_bbox[5])
-            for i in range(len(item_bbox) // 2):
-                if item_bbox[i] < collection_bbox[i]:
-                    collection_bbox[i] = item_bbox[i]
-            for i in range(len(item_bbox) // 2, len(item_bbox)):
-                if item_bbox[i] > collection_bbox[i]:
-                    collection_bbox[i] = item_bbox[i]
-        elif item_bbox:
-            collection_bboxes.append(item_bbox)
-        self._check_wgs84_compliance(collection_bboxes[0], "collection", self._collection.get("id"))
-
-    def _update_collection_interval(self, stac_item: dict[str, Any]) -> None:
-        if (datetime := stac_item["properties"].get("datetime")) is not None:
-            item_interval = [datetime, datetime]
-        else:
-            item_interval = [stac_item["properties"][prop] for prop in ("start_datetime", "end_datetime")]
-        collection_intervals = self._collection["extent"]["temporal"]["interval"]
-        if collection_intervals:
-            collection_interval = collection_intervals[0]
-            if collection_interval[0] is not None and item_interval[0] < collection_interval[0]:
-                collection_interval[0] = item_interval[0]
-            if collection_interval[1] is not None and item_interval[1] > collection_interval[1]:
-                collection_interval[1] = item_interval[1]
-        else:
-            collection_intervals.append(item_interval)
-
-    def _update_collection_summaries(self, stac_item: dict[str, Any]) -> None:
-        """
-        Update the summaries value in the stac_collection based on the values in stac_item.
-
-        This only creates summaries for simple types (strings, numbers, boolean) and does not
-        create summaries as JSON schema objects.
-        """
-        if "summaries" not in self._collection:
-            self._collection["summaries"] = {}
-        elif "needs_summaries_update" in self._collection["summaries"]:
-            self._collection["summaries"].pop("needs_summaries_update")
-        summaries = self._collection["summaries"]
-        for name, value in stac_item["properties"].items():
-            summary = summaries.get(name)
-            if name in self.exclude_summaries:
-                continue
-            elif isinstance(value, bool):
-                if summary is None:
-                    summaries[name] = [value]
-                elif value not in summary:
-                    summary.append(value)
-            elif isinstance(value, str):
-                try:
-                    time_value = datetime.fromisoformat(value)
-                except ValueError:
-                    if summary is None:
-                        summaries[name] = [value]
-                    elif isinstance(summary, list):
-                        if value not in summary:
-                            summary.append(value)
-                else:
-                    if summary is None:
-                        summaries[name] = {"minimum": value, "maximum": value}
-                    elif summary.get("minimum") is not None and summary.get("maximum") is not None:
-                        if time_value < datetime.fromisoformat(summary["minimum"]):
-                            summary["minimum"] = value
-                        elif time_value > datetime.fromisoformat(summary["maximum"]):
-                            summary["maximum"] = value
-            elif isinstance(value, (int, float)):
-                if summary is None:
-                    summaries[name] = {"minimum": value, "maximum": value}
-                elif isinstance(summary, list):
-                    # this property does not necessarily contain all numeric values
-                    if value not in summary:
-                        summary.append(value)
-                elif summary.get("minimum") is not None and summary.get("maximum") is not None:
-                    if value < summary["minimum"]:
-                        summary["minimum"] = value
-                    elif value > summary["maximum"]:
-                        summary["maximum"] = value
-
     def _update_collection(self, stac_item: dict[str, Any]) -> None:
-        if self.update:
-            if self.update_collection in ("extents", "all"):
-                LOGGER.info(
-                    "Updating collection extents [%s] with data from item [%s]",
-                    self._collection.get("id"),
-                    stac_item.get("id"),
-                )
-                self._update_collection_bbox(stac_item)
-                self._update_collection_interval(stac_item)
-            if self.update_collection in ("summaries", "all"):
-                LOGGER.info(
-                    "Updating collection summaries [%s] with data from item [%s]",
-                    self._collection.get("id"),
-                    stac_item.get("id"),
-                )
-                self._update_collection_summaries(stac_item)
+        if self.update and self.update_collection != "none":
+            update_collection(self.update_collection, self._collection, stac_item, self.exclude_summaries)
 
     def ingest(self) -> None:
         """Ingest data."""